// shared/enums/category.enums.ts
/** @summary Supported category types. */
export enum CategoryType {
    INCOME = "income",
    EXPENSE = "expense",
}

/** @summary Supported category colors. */
export enum CategoryColor {
    RED = "red",
    BLUE = "blue",
    GREEN = "green",
    PURPLE = "purple",
    YELLOW = "yellow",
    ORANGE = "orange",
    PINK = "pink",
    GRAY = "gray",
    CYAN = "cyan",
    INDIGO = "indigo",
}